backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
pillow